requirements.txt
pymongo==3.0.3
python-amazon-product-api==0.2.8
tornado==4.0.2
